requirements.txt
aiohttp==3.8.4
aiohttp-jinja2==1.5.1
gunicorn==20.1.0
tortoise-orm==0.19.3
asyncpg==0.27.0
python-dotenv==1.0.0
cryptography==39.0.2
PyJWT==2.6.0
